cleanup: use same export mechanism for all modules
Cleaned up the way we are exporting functionality from this crate.
Only export the functionality using public modules because these
modules logically separate the functionality and do not make sense
as a bulk.

As part of this cleanup, also moved fallocate to its own module
allong with the helper structure.

As a side effect of the cleanup, the main documentation page of the
crate is much simpler now and easier to navigate through.
